FIX MLX WINDOW POPING IF ERRORS:
Now checking the parsing errors first then init mlx window.
